move files into /cmd
